+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Remove caret entry direction when selecting parent of caret.
 -   Only highlight definitions and uses of caret position when inside a token.
 -   Fixed incorrect placement of caret; it was assuming pretty printing.
+-   Removed unused font preload.
 
 ### Maintenance